% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_browser.R
\name{connect_to_browser}
\alias{connect_to_browser}
\title{setup chrome or edge broser dirver and selenium server}
\usage{
connect_to_browser(browser = "chrome", download_path)
}
\arguments{
\item{browser}{The browser you can choose chrome or edge, default is chrome}

\item{download_path}{The path of chrome or edge broser dirver and selenium server}
}
\description{
\code{connect_to_browser()} is deprecated. We will soon be totally
}
\examples{
connect_to_browser(browser = "chrome", download_path = "C:/Users/Dell/Downloads/selenium")
connect_to_browser(browser = "edge", download_path = "C:/Users/Dell/Downloads/selenium")
}
