% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_from_browser.R
\name{disconnect_from_browser}
\alias{disconnect_from_browser}
\title{stop chrome or edge broser dirver and selenium server}
\usage{
disconnect_from_browser()
}
\description{
\code{disconnect_from_browser()} is deprecated. We will soon be totally
}
\examples{
disconnect_from_browser()
}
