% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_by_hand.R
\name{download_by_hand}
\alias{download_by_hand}
\title{Add one to any value}
\usage{
download_by_hand(location, data_date, url, username, password, ...)
}
\arguments{
\item{location}{A numeric value}

\item{data_date}{The date of data download}

\item{url}{The url of pig performance test station website}

\item{username}{The username of pig performance test station website}

\item{password}{The password of pig performance test station website}

\item{...}{other parameters}
}
\value{
csv datas
}
\description{
Add one to any value
}
\examples{
web_ppt1 <- "https://yxlxppt01.vpu-online.cn/login/LoginPage.web"
username1 <- "隆鑫场测定一区"
password1 <- "yangxianglongxin1234"
station_create5 <- purrr::map_chr(seq(101L,102L,1L), c)
download_by_hand(location = station_create5, 
                 data_date = "2023-02-11", 
                 url = web_ppt1, 
                 username = username1, 
                 password = password1)

}
