% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_schedule.R
\name{download_csv_schedule}
\alias{download_csv_schedule}
\title{Schedule automatic downloads of CSVs}
\usage{
download_csv_schedule(
  taskname,
  schedule,
  starttime,
  startdate,
  rscript_args = NULL,
  ...
)
}
\arguments{
\item{taskname}{A character string with the name of the task. Defaults to the filename. Should not contain any spaces.}

\item{schedule}{Either one of 'ONCE', 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE', 'ONLOGON', 'ONIDLE.}

\item{starttime}{A time point in HH:mm format indicating when to run the script. Defaults to within 62 seconds.}

\item{startdate}{A date that specifies the first date on which to run the task. Only applicable if schedule is of type 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE'. Defaults to today in '\%d/\%m/\%Y' format. Change to your locale format if needed.}

\item{rscript_args}{Character string with further arguments passed on to Rscript.}

\item{...}{other parameters.}
}
\value{
CSVs in the path of save_path.
}
\description{
\code{download_csv_schedule()} is used to make download task.
}
\examples{
# Make a download task
download_csv_schedule(taskname = "ppt_csv_download",
                      schedule = "DAILY",
                      starttime = "20:00",
                      startdate = format(Sys.Date(), "\%Y/\%m/\%d"),
                      rscript_args = list(browser = "chrome", 
                                          url = Sys.getenv("web_url"),
                                          username = Sys.getenv("user_name"),
                                          password = Sys.getenv("user_password"),
                                          location = c("606","607"),
                                          csv_position = "1",
                                          download_path = "C:/Users/Dell/Downloads/download_temp",
                                          save_path = "C:/Users/Dell/Downloads/download_rename"))
# Delete download task
taskscheduleR::taskscheduler_delete("ppt_csv_download")
# Reset selenium webdriver service
disconnect_from_browser()
}
