% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_free_brain.R
\name{download_free_brain}
\alias{download_free_brain}
\title{download free brain}
\usage{
download_free_brain(location, data_date, ...)
}
\arguments{
\item{location}{Numeric value in string format}

\item{data_date}{The dates of data download}

\item{...}{other parameters}
}
\value{
CSV files in the download path of your Google Chrome web browser.
}
\description{
\code{download_free_brain()} is deprecated. We will soon be totally
}
\examples{
user_name = "隆鑫场测定一区"
stations = purrr::map_chr(seq(101L,103L,1L), c)
download_date = seq(as.Date("2023-02-01"),as.Date("2023-02-03"),1)
download_free_brain(location = stations, 
                    data_date = download_date, 
                    url = Sys.getenv("web_url"), 
                    username = user_name, 
                    password = Sys.getenv("user_password"))
}
