% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_from_browser.R
\name{disconnect_from_browser}
\alias{disconnect_from_browser}
\title{Reset selenium webdriver service}
\usage{
disconnect_from_browser()
}
\description{
\code{disconnect_from_browser()} is used to reset selenium webdriver service.
}
\examples{
disconnect_from_browser()
}
