% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_by_hand.R
\name{download_by_hand}
\alias{download_by_hand}
\title{Add one to any value}
\usage{
download_by_hand(location, data_date, url, username, password, ...)
}
\arguments{
\item{location}{A numeric value}

\item{data_date}{The date of data download}

\item{url}{The url of pig performance test station website}

\item{username}{The username of pig performance test station website}

\item{password}{The password of pig performance test station website}

\item{...}{other parameters}
}
\value{
csv datas
}
\description{
Add one to any value
}
\examples{
download_by_hand(location = Sys.getenv("stations"), 
                 data_date = Sys.getenv("download_date"), 
                 url = Sys.getenv("web_url"),
                 username = Sys.getenv("user_name"), 
                 password = Sys.getenv("user_password"))

}
