% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_auto.R
\name{download_csv_auto}
\alias{download_csv_auto}
\title{Automatically download CSVs}
\usage{
download_csv_auto(
  url,
  username,
  password,
  csv_position,
  location,
  date,
  download_path
)
}
\arguments{
\item{url}{The login url of pig performance test station website.}

\item{username}{The username of pig performance test station website.}

\item{password}{The password of pig performance test station website.}

\item{csv_position}{The pisiton of csv download item.}

\item{location}{Numeric value in string format.}

\item{date}{The dates of data download.}

\item{download_path}{The folder of downloading csv files.}
}
\value{
CSVs in the path of save_path.
}
\description{
\code{download_csv_auto()} is used to download CSVs from chrome browser.
}
\examples{
# Start selenium webdriver service
connect_to_browser(dest_dir = "C:/Users/Dell/Desktop/test")
# Download CSVs
download_csv_auto(url = Sys.getenv("url"), 
                  username = Sys.getenv("username"), 
                  password = Sys.getenv("password"),
                  csv_position = "1",
                  location = c("606","607"), 
                  date = c("2024-02-18","2024-02-19"), 
                  download_path = "C:/Users/Dell/Downloads/download_rename")
}
