% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_auto.R
\name{download_csv_auto}
\alias{download_csv_auto}
\title{download ppt csv data acorrding the way you want}
\usage{
download_csv_auto(
  url,
  username,
  password,
  csv_position,
  location,
  data_date,
  download_path,
  save_path,
  ...
)
}
\arguments{
\item{url}{The url of pig performance test station website}

\item{username}{The username of pig performance test station website}

\item{password}{The password of pig performance test station website}

\item{csv_position}{The pisiton of csv download item}

\item{location}{Numeric value in string format}

\item{data_date}{The dates of data download}

\item{download_path}{The default folder of Google Chrome web browser}

\item{save_path}{The folder you need to save renamed csv files}

\item{...}{other parameters}
}
\value{
CSV files in the download path of your Google Chrome web browser.
}
\description{
\code{download_csv_auto} is deprecated. We will soon be totally
}
\examples{
stations = c("1801","1802")
data_date = c("2023-03-29","2023-03-30")
download_path = "C:/Users/Dell/Downloads/"
save_path = "C:/Users/Dell/Downloads/test/"

download_csv_auto(url = Sys.getenv("web_url"), 
                  username = Sys.getenv("user_name"), 
                  password = Sys.getenv("user_password"),
                  csv_position = "1",
                  location = stations, 
                  data_date = data_date, 
                  download_path = download_path,
                  save_path = save_path)

}
