% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_auto.R
\name{download_csv_auto}
\alias{download_csv_auto}
\title{Automatically download CSV}
\usage{
download_csv_auto(
  browser,
  url,
  username,
  password,
  csv_position,
  location,
  data_date,
  download_path,
  save_path,
  ...
)
}
\arguments{
\item{browser}{The browser you can chrome or edge}

\item{url}{The url of pig performance test station website}

\item{username}{The username of pig performance test station website}

\item{password}{The password of pig performance test station website}

\item{csv_position}{The pisiton of csv download item}

\item{location}{Numeric value in string format}

\item{data_date}{The dates of data download}

\item{download_path}{The folder of downloading csv files}

\item{save_path}{The folder you need to save renamed csv files}

\item{...}{other parameters}
}
\value{
CSV files in the path of save_path
}
\description{
\code{download_csv_auto()} is deprecated. We will soon be totally
}
