% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_schedule.R
\name{download_csv_schedule}
\alias{download_csv_schedule}
\title{Schedule automatic downloads of CSV}
\usage{
download_csv_schedule(
  taskname,
  schedule,
  starttime,
  startdate,
  rscript_args,
  ...
)
}
\arguments{
\item{taskname}{A character string with the name of the task. Defaults to the filename. Should not contain any spaces}

\item{schedule}{Either one of 'ONCE', 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE', 'ONLOGON', 'ONIDLE}

\item{starttime}{A timepoint in HH:mm format indicating when to run the script. Defaults to within 62 seconds}

\item{startdate}{A date that specifies the first date on which to run the task. Only applicable if schedule is of type 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE'. Defaults to today in '\%d/\%m/\%Y' format. Change to your locale format if needed}

\item{rscript_args}{Character string with further arguments passed on to Rscript}

\item{...}{other parameters}
}
\value{
CSV files in the path of save_path
}
\description{
\code{download_csv_schedule()} is deprecated. We will soon be totally
}
\examples{
download_csv_schedule(taskname = "ppt_csv_download",
                      schedule = "DAILY",
                      starttime = "20:05",
                      startdate = format(Sys.Date(), "\%Y/\%m/\%d"),
                      rscript_args = list(url = Sys.getenv("web_url"),
                                          username = Sys.getenv("user_name"),
                                          password = Sys.getenv("user_password"),
                                          location = c("1801","1802"),
                                          data_date = Sys.Date()-1,
                                          csv_position = "1",
                                          download_path = "C:/Users/Dell/Downloads",
                                          save_path = "C:/Users/Dell/Downloads/test"))
}
