% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_free_hands.R
\name{download_free_hands}
\alias{download_free_hands}
\title{download free hands}
\usage{
download_free_hands(
  location,
  data_date,
  url,
  username,
  password,
  csv_position,
  ...
)
}
\arguments{
\item{location}{Numeric value in string format}

\item{data_date}{The dates of data download}

\item{url}{The url of pig performance test station website}

\item{username}{The username of pig performance test station website}

\item{password}{The password of pig performance test station website}

\item{...}{other parameters}
}
\value{
CSV files in the download path of your Google Chrome web browser.
}
\description{
\code{download_free_hands()} is deprecated. We will soon be totally
}
\examples{
user_name = "隆鑫场测定一区"
stations = "101"
download_date = "2023-02-11"
download_free_hands(location = stations, 
                    data_date = download_date, 
                    url = Sys.getenv("web_url"), 
                    username = user_name, 
                    password = Sys.getenv("user_password"))

}
