% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv.R
\name{get_csv}
\alias{get_csv}
\title{stop chrome or edge broser dirver and selenium server}
\usage{
get_csv(
  url,
  username,
  password,
  location_min,
  location_max,
  date_min,
  date_max,
  download_path,
  show_message = FALSE
)
}
\arguments{
\item{url}{A character string specifying the URL to log into.}

\item{username}{A character string specifying the username for login.}

\item{password}{A character string specifying the password for login.}

\item{location_min}{A character string specifying the minimum location.}

\item{location_max}{A character string specifying the maximum location.}

\item{date_min}{A Date object specifying the earliest date.}

\item{date_max}{A Date object specifying the latest date.}

\item{download_path}{A character string specifying the path where the downloaded CSV file should be saved.}

\item{show_message}{A logical value. If TRUE, a message will be displayed after the download. Default is FALSE.}
}
\value{
This function has no return value. It downloads a CSV file and saves it to the specified path.
}
\description{
This function logs into a given URL with a username and password, and downloads a CSV file.
}
\examples{
get_csv(url = Sys.getenv("url"),
        username = Sys.getenv("username"),
        password = Sys.getenv("password"),
        location_min = "406",
        location_max = "407",
        date_min = "2024-02-18",
        date_max = "2024-02-19",
        download_path = "C:/Users/Dell/Downloads/download_rename", 
        show_message = TRUE)
}
