% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv_auto.R
\name{get_csv_auto}
\alias{get_csv_auto}
\title{stop chrome or edge broser dirver and selenium server}
\usage{
get_csv_auto(
  url,
  username,
  password,
  location,
  date,
  download_path,
  show_message = FALSE
)
}
\arguments{
\item{url}{A character string specifying the URL to log into.}

\item{username}{A character string specifying the username for login.}

\item{password}{A character string specifying the password for login.}

\item{location}{A character string specifying the location.}

\item{date}{A Date object specifying the date.}

\item{download_path}{A character string specifying the path where the downloaded CSV file should be saved.}

\item{show_message}{A logical value. If TRUE, a message will be displayed after the download. Default is FALSE.}
}
\value{
This function has no return value. It downloads a CSV file and saves it to the specified path.
}
\description{
This function logs into a given URL with a username and password, and downloads a CSV file.
}
\examples{
get_csv_auto(url = Sys.getenv("url"),
             username = Sys.getenv("username"),
             password = Sys.getenv("password"),
             location = c("607","608"),
             date = c("2024-02-18", "2024-02-19"),
             download_path = "C:/Users/Dell/Downloads/download_rename")
}
